Re-organize client directory by moving related docs into separate directories. #4
