+++ client/episode/index.js
@@ -1,0 +1,3 @@
+import EpisodeContainer from './episode-container'
+
+export default EpisodeContainer
