+++ client/subscription/index.js
@@ -1,0 +1,3 @@
+import SubContainer from './sub-container'
+
+export default SubContainer
