+++ client/subscription/sub-cover-art.js
